+++ lettuce/components/embeddings.py
@@ -1,11 +1,9 @@
-from enum import Enum
 from haystack_integrations.components.embedders.fastembed import (
     FastembedTextEmbedder,
 )
 from haystack import component
 from haystack.dataclasses import Document
 from typing import Any, List, Dict
-from pydantic import BaseModel
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm import Session
 
@@ -13,89 +11,12 @@ from omop.omop_queries import query_vector
 from omop.db_manager import db_session
 
 from options.base_options import BaseOptions
+from options.pipeline_options import EmbeddingModelName, EmbeddingModel, EMBEDDING_MODELS
 
 settings = BaseOptions()
 
 # -------- Embedding Models -------- >
 
-
-class EmbeddingModelName(str, Enum):
-    """
-    This class enumerates the embedding models we
-    have the download details for.
-
-    The models are:
-    """
-
-    BGESMALL = "BGESMALL"
-    MINILM = "MINILM"
-    GTR_T5_BASE = "gtr-t5-base"
-    GTR_T5_LARGE = "gtr-t5-large"
-    E5_BASE = "e5-base"
-    E5_LARGE = "e5-large"
-    DISTILBERT_BASE_UNCASED = "distilbert-base-uncased"
-    DISTILUSE_BASE_MULTILINGUAL = "distiluse-base-multilingual-cased-v1"
-    CONTRIEVER = "contriever"
-
-
-class EmbeddingModelInfo(BaseModel):
-    """
-    A simple class to hold the information for embeddings models
-    """
-
-    path: str
-    dimensions: int
-
-
-class EmbeddingModel(BaseModel):
-    """
-    A class to match the name of an embeddings model with the
-    details required to download and use it.
-    """
-
-    name: EmbeddingModelName
-    info: EmbeddingModelInfo
-
-
-EMBEDDING_MODELS = {
-    # ------ Bidirectional Gated Encoder  ------- >
-    EmbeddingModelName.BGESMALL: EmbeddingModelInfo(
-        path="BAAI/bge-small-en-v1.5", dimensions=384
-    ),
-    # ------ SBERT (Sentence-BERT) ------- >
-    EmbeddingModelName.MINILM: EmbeddingModelInfo(
-        path="sentence-transformers/all-MiniLM-L6-v2", dimensions=384
-    ),
-    # ------ Generalizable T5 Retrieval ------- >
-    EmbeddingModelName.GTR_T5_BASE: EmbeddingModelInfo(
-        path="google/gtr-t5-base", dimensions=768
-    ),
-    # ------ Generalizable T5 Retrieval ------- >
-    EmbeddingModelName.GTR_T5_LARGE: EmbeddingModelInfo(
-    path="google/gtr-t5-large", dimensions=1024
-    ),
-    # ------ Embedding Models for Search Engines ------- >
-    EmbeddingModelName.E5_BASE: EmbeddingModelInfo(
-        path="microsoft/e5-base", dimensions=768
-    ),
-    # ------ Embedding Models for Search Engines ------- >
-    EmbeddingModelName.E5_LARGE: EmbeddingModelInfo(
-        path="microsoft/e5-large", dimensions=1024
-    ),
-    # ------ DistilBERT ------- >
-    EmbeddingModelName.DISTILBERT_BASE_UNCASED: EmbeddingModelInfo(
-        path="distilbert-base-uncased", dimensions=768
-    ),
-    # ------ distiluse-base-multilingual-cased-v1 ------- >
-    EmbeddingModelName.DISTILUSE_BASE_MULTILINGUAL: EmbeddingModelInfo(
-        path="sentence-transformers/distiluse-base-multilingual-cased-v1",
-        dimensions=512,
-    ),
-    # ------ Contriever ------- >
-    EmbeddingModelName.CONTRIEVER: EmbeddingModelInfo(
-        path="facebook/contriever", dimensions=768
-    ),
-}
 
 # I know there's a pgvector integration with haystack, but it seems to be able to only query a single table at a time
 # https://github.com/deepset-ai/haystack-core-integrations/tree/main/integrations/pgvector/examples
